android: Avoid misleading serviceLost/serviceFound events
The current implementation of `queryTxtRecord` in the Android-specific
Bonsoir event service triggers `serviceLost` and `serviceFound` events
when attempting to read TXT records immediately after discovering a
service. This occurs because Android's NSD does not support reading
TXT records without first resolving the service.

For applications that do not rely on TXT records before resolving the
service, this behavior is unnecessary and misleading. This commit
removes the `queryTxtRecord` logic to ensure that `serviceLost` and
`serviceFound` events correctly reflect the actual state of the MDNS
network rather than being artifacts of the implementation.

Signed-off-by: Daniel Jenikovský <daniel.jenikovsky@streamunlimited.com>