ios: Do not broadcast serviceLost and serviceFound events on change
Service change events previously broadcasted pair of serviceLost/Found
events to notify the application about change. This however doesn't
represent the real state of the network and causes issues in the app.

Because our app doesn't rely on change events, they can be removed
completely. This way, only real MDNS events are reported to the
application.

Signed-off-by: Daniel Jenikovský <daniel.jenikovsky@streamunlimited.com>